refactor(render): split dir_model.go into focused modules
Split the monolithic render/dir_model.go into smaller files grouped by
responsibility: messages, language filtering, sorting, table data,
treemap, chart, global search, and mouse handling. Move related tests
into matching files and clean up a few minor style issues.
