+++ render/dir_model_test.go
@@ -7,8 +7,6 @@ import (
 	"testing"
 
 	tea "github.com/charmbracelet/bubbletea"
-	"github.com/charmbracelet/lipgloss"
-	"github.com/stretchr/testify/require"
 
 	"github.com/zdyxry/tokui/filter"
 	"github.com/zdyxry/tokui/provider"
@@ -26,28 +24,6 @@ func TestColumnFmtName(t *testing.T) {
 	}
 	if got := col.FmtName(SortState{Key: SortByTotal, Desc: true}); got != "Total ▼" {
 		t.Errorf("expected 'Total ▼', got %q", got)
-	}
-}
-
-func TestSortStateDirectionArrow(t *testing.T) {
-	if got := (SortState{Desc: false}).DirectionArrow(); got != "▲" {
-		t.Errorf("expected '▲', got %q", got)
-	}
-	if got := (SortState{Desc: true}).DirectionArrow(); got != "▼" {
-		t.Errorf("expected '▼', got %q", got)
-	}
-}
-
-func TestDefaultDescForSortKey(t *testing.T) {
-	for _, key := range []SortKey{SortByCode, SortByComments, SortByBlanks, SortByTotal, SortByPercent} {
-		if !defaultDescForSortKey(key) {
-			t.Errorf("expected numeric key %q to default to descending", key)
-		}
-	}
-	for _, key := range []SortKey{SortByName, SortByLanguages} {
-		if defaultDescForSortKey(key) {
-			t.Errorf("expected text key %q to default to ascending", key)
-		}
 	}
 }
 
@@ -79,221 +55,6 @@ func newTestDirModel() *DirModel {
 	return dm
 }
 
-func TestDirModelComparableStats(t *testing.T) {
-	dm := newTestDirModel()
-	root := dm.nav.Entry()
-
-	t.Run("no filter uses total stats", func(t *testing.T) {
-		got := dm.comparableStats(root.Child[0])
-		want := structure.CodeStats{Code: 20, Comments: 5, Blanks: 5}
-		if got != want {
-			t.Errorf("a.go total stats: got %+v, want %+v", got, want)
-		}
-	})
-
-	t.Run("single language filter", func(t *testing.T) {
-		dm.langFilterIdx = 0                     // Go
-		got := dm.comparableStats(root.Child[1]) // b.py has no Go stats
-		want := structure.CodeStats{}
-		if got != want {
-			t.Errorf("b.py Go stats: got %+v, want %+v", got, want)
-		}
-	})
-
-	t.Run("multi language filter aggregates", func(t *testing.T) {
-		dm.langFilterIdx = -1
-		dm.selectedLangs["Go"] = true
-		dm.selectedLangs["Python"] = true
-		got := dm.comparableStats(root)
-		want := structure.CodeStats{Code: 60, Comments: 17, Blanks: 18}
-		if got != want {
-			t.Errorf("root aggregated stats: got %+v, want %+v", got, want)
-		}
-	})
-}
-
-func TestDirModelBuildChildComparator(t *testing.T) {
-	t.Run("sort by name ascending", func(t *testing.T) {
-		dm := newTestDirModel()
-		dm.sortState = SortState{Key: SortByName, Desc: false}
-		dm.nav.Entry().SortChildBy(dm.buildChildComparator())
-		want := []string{"a.go", "b.py", "c.go"}
-		for i, child := range dm.nav.Entry().Child {
-			if child.Name() != want[i] {
-				t.Errorf("position %d: expected %q, got %q", i, want[i], child.Name())
-			}
-		}
-	})
-
-	t.Run("sort by code descending", func(t *testing.T) {
-		dm := newTestDirModel()
-		dm.sortState = SortState{Key: SortByCode, Desc: true}
-		dm.nav.Entry().SortChildBy(dm.buildChildComparator())
-		want := []int64{30, 20, 10}
-		for i, child := range dm.nav.Entry().Child {
-			got := dm.comparableStats(child).Code
-			if got != want[i] {
-				t.Errorf("position %d: expected code %d, got %d", i, want[i], got)
-			}
-		}
-	})
-
-	t.Run("sort by total ascending", func(t *testing.T) {
-		dm := newTestDirModel()
-		dm.sortState = SortState{Key: SortByTotal, Desc: false}
-		dm.nav.Entry().SortChildBy(dm.buildChildComparator())
-		want := []int64{15, 30, 50}
-		for i, child := range dm.nav.Entry().Child {
-			got := dm.comparableStats(child).Total()
-			if got != want[i] {
-				t.Errorf("position %d: expected total %d, got %d", i, want[i], got)
-			}
-		}
-	})
-
-	t.Run("sort by percent descending", func(t *testing.T) {
-		dm := newTestDirModel()
-		dm.sortState = SortState{Key: SortByPercent, Desc: true}
-		dm.nav.Entry().SortChildBy(dm.buildChildComparator())
-		want := []int64{50, 30, 15}
-		for i, child := range dm.nav.Entry().Child {
-			got := dm.comparableStats(child).Total()
-			if got != want[i] {
-				t.Errorf("position %d: expected total %d, got %d", i, want[i], got)
-			}
-		}
-	})
-
-	t.Run("multi-language sort by code descending", func(t *testing.T) {
-		dm := newTestDirModel()
-		dm.selectedLangs["Go"] = true
-		dm.selectedLangs["Python"] = true
-		dm.sortState = SortState{Key: SortByCode, Desc: true}
-		dm.nav.Entry().SortChildBy(dm.buildChildComparator())
-		// a.go=20, c.go=30, b.py=10 when aggregated
-		want := []int64{30, 20, 10}
-		for i, child := range dm.nav.Entry().Child {
-			got := dm.comparableStats(child).Code
-			if got != want[i] {
-				t.Errorf("position %d: expected code %d, got %d", i, want[i], got)
-			}
-		}
-	})
-}
-
-func TestDirModelCycleSortColumn(t *testing.T) {
-	dm := newTestDirModel()
-	// newTestDirModel initializes sortState to SortByTotal, so the first cycle
-	// Starting from SortByTotal, the cycle is Percent, Complexity,
-	// Name, Languages, Code, Comments, Blanks, then wraps back to Total.
-	order := []SortKey{SortByPercent, SortByComplexity, SortByName, SortByLanguages, SortByCode, SortByComments, SortByBlanks, SortByTotal}
-
-	for i := 0; i < len(order)*2; i++ {
-		expected := order[i%len(order)]
-		dm.cycleSortColumn()
-		if dm.sortState.Key != expected {
-			t.Errorf("cycle %d: expected key %q, got %q", i, expected, dm.sortState.Key)
-		}
-		if dm.sortState.Desc != defaultDescForSortKey(expected) {
-			t.Errorf("cycle %d: expected default desc %v for %q", i, defaultDescForSortKey(expected), expected)
-		}
-	}
-}
-
-func TestDirModelToggleSortOrder(t *testing.T) {
-	dm := newTestDirModel()
-	dm.sortState = SortState{Key: SortByTotal, Desc: true}
-
-	dm.toggleSortOrder()
-	if dm.sortState.Desc != false {
-		t.Errorf("expected Desc=false after toggle")
-	}
-	dm.toggleSortOrder()
-	if dm.sortState.Desc != true {
-		t.Errorf("expected Desc=true after second toggle")
-	}
-	if dm.sortState.Key != SortByTotal {
-		t.Errorf("expected key to remain %q", SortByTotal)
-	}
-}
-
-func TestDirModelGlobalSearchOpenClose(t *testing.T) {
-	dm := newTestDirModel()
-	dm.Update(ScanFinished{})
-
-	dm.openGlobalSearch()
-	if dm.mode != SEARCH {
-		t.Fatalf("expected SEARCH mode, got %v", dm.mode)
-	}
-	if dm.searchIndex == nil {
-		t.Fatal("expected search index to be built")
-	}
-
-	dm.closeGlobalSearch()
-	if dm.mode != READY {
-		t.Fatalf("expected READY mode after close, got %v", dm.mode)
-	}
-}
-
-func TestDirModelGlobalSearchFindsFile(t *testing.T) {
-	dm := newTestDirModel()
-	dm.Update(ScanFinished{})
-	dm.openGlobalSearch()
-
-	dm.searchInput.SetValue("b.py")
-	dm.updateSearchQuery()
-
-	if len(dm.searchMatches) == 0 {
-		t.Fatalf("expected matches for 'b.py', got none")
-	}
-	if dm.searchMatches[0].Item.Path != "b.py" {
-		t.Errorf("expected first match to be b.py, got %q", dm.searchMatches[0].Item.Path)
-	}
-}
-
-func TestDirModelApplySearchResultDirectory(t *testing.T) {
-	dm := newTestDirModel()
-	dm.Update(ScanFinished{})
-	dm.openGlobalSearch()
-
-	// "root" is the top-level directory; selecting it should keep us at root.
-	dm.searchInput.SetValue("root")
-	dm.updateSearchQuery()
-	dm.applySearchResult()
-
-	if dm.mode != READY {
-		t.Fatalf("expected READY mode after apply, got %v", dm.mode)
-	}
-	if dm.nav.Entry().Path != "root" {
-		t.Errorf("expected nav at root, got %q", dm.nav.Entry().Path)
-	}
-}
-
-func TestDirModelApplySearchResultFile(t *testing.T) {
-	dm := newTestDirModel()
-	dm.Update(ScanFinished{})
-	dm.openGlobalSearch()
-
-	dm.searchInput.SetValue("b.py")
-	dm.updateSearchQuery()
-	dm.applySearchResult()
-
-	if dm.mode != READY {
-		t.Fatalf("expected READY mode after apply, got %v", dm.mode)
-	}
-	if dm.nav.Entry().Path != "root" {
-		t.Errorf("expected nav at parent directory root, got %q", dm.nav.Entry().Path)
-	}
-
-	idx := dm.findChildIndex(dm.nav.Entry().GetChild("b.py"))
-	if idx < 0 {
-		t.Fatal("expected b.py to be visible in the table")
-	}
-	if dm.dirsTable.Cursor() != idx {
-		t.Errorf("expected cursor at b.py index %d, got %d", idx, dm.dirsTable.Cursor())
-	}
-}
-
 func newTestNestedDirModel() *DirModel {
 	root := structure.NewDirEntry("root")
 
@@ -315,174 +76,6 @@ func newTestNestedDirModel() *DirModel {
 	dm.selectedLangs = make(map[string]bool)
 	dm.sortState = SortState{Key: SortByTotal, Desc: true}
 	return dm
-}
-
-func TestDirModelApplySearchResultFromTreeMode(t *testing.T) {
-	dm := newTestNestedDirModel()
-	dm.Update(ScanFinished{})
-	dm.treeMode = true
-	dm.openGlobalSearch()
-
-	dm.searchInput.SetValue("b.py")
-	dm.updateSearchQuery()
-	dm.applySearchResult()
-
-	if !dm.treeMode {
-		t.Error("expected tree mode to remain active after global search jump")
-	}
-	if dm.treemapMode {
-		t.Error("expected treemap mode to remain disabled after global search jump")
-	}
-	if dm.nav.Entry().Path != "root" {
-		t.Errorf("expected tree root to stay at project root, got %q", dm.nav.Entry().Path)
-	}
-	subdir := dm.nav.Entry().GetChild("subdir")
-	require.NotNil(t, subdir, "expected subdir to be visible under root")
-	if !subdir.Expanded {
-		t.Error("expected subdir to be expanded so the target file is visible")
-	}
-
-	target := subdir.GetChild("b.py")
-	require.NotNil(t, target, "expected b.py to be visible under subdir")
-	idx := dm.findChildIndex(target)
-	if idx < 0 {
-		t.Fatal("expected b.py to be visible in the tree table")
-	}
-	if dm.dirsTable.Cursor() != idx {
-		t.Errorf("expected cursor on b.py index %d, got %d", idx, dm.dirsTable.Cursor())
-	}
-}
-
-func TestDirModelApplySearchResultFromTreeModeDirectory(t *testing.T) {
-	dm := newTestNestedDirModel()
-	dm.Update(ScanFinished{})
-	dm.treeMode = true
-	dm.openGlobalSearch()
-
-	dm.searchInput.SetValue("subdir")
-	dm.updateSearchQuery()
-	dm.applySearchResult()
-
-	if !dm.treeMode {
-		t.Error("expected tree mode to remain active after global search jump")
-	}
-	if dm.nav.Entry().Path != "root" {
-		t.Errorf("expected nav at parent directory root, got %q", dm.nav.Entry().Path)
-	}
-	if !dm.nav.Entry().GetChild("subdir").Expanded {
-		t.Error("expected subdir to be expanded after searching for it in tree mode")
-	}
-
-	idx := dm.findChildIndex(dm.nav.Entry().GetChild("subdir"))
-	if idx < 0 {
-		t.Fatal("expected subdir to be visible after search jump")
-	}
-	if dm.dirsTable.Cursor() != idx {
-		t.Errorf("expected cursor on subdir index %d, got %d", idx, dm.dirsTable.Cursor())
-	}
-}
-
-func TestDirModelApplySearchResultFromTreemapMode(t *testing.T) {
-	dm := newTestNestedDirModel()
-	dm.Update(ScanFinished{})
-	dm.width = 80
-	dm.height = 24
-	dm.treemapMode = true
-	dm.openGlobalSearch()
-
-	dm.searchInput.SetValue("b.py")
-	dm.updateSearchQuery()
-	dm.applySearchResult()
-
-	if !dm.treemapMode {
-		t.Error("expected treemap mode to remain active after global search jump")
-	}
-	if dm.treeMode {
-		t.Error("expected tree mode to remain disabled after global search jump")
-	}
-	if dm.nav.Entry().Path != "root/subdir" {
-		t.Errorf("expected nav at parent directory root/subdir, got %q", dm.nav.Entry().Path)
-	}
-
-	_ = dm.View() // trigger treemap block layout
-
-	idx := dm.findTreemapBlockIndex(dm.nav.Entry().GetChild("b.py"))
-	if idx < 0 {
-		t.Fatalf("expected b.py to be visible in treemap blocks after search jump, got %d blocks", len(dm.treemapBlocks))
-	}
-	if dm.treemapSelected != idx {
-		t.Errorf("expected treemap selection on b.py index %d, got %d", idx, dm.treemapSelected)
-	}
-}
-
-func TestDirModelApplySearchResultFromTreemapModeDirectory(t *testing.T) {
-	dm := newTestNestedDirModel()
-	dm.Update(ScanFinished{})
-	dm.width = 80
-	dm.height = 24
-	dm.treemapMode = true
-	dm.openGlobalSearch()
-
-	dm.searchInput.SetValue("subdir")
-	dm.updateSearchQuery()
-	dm.applySearchResult()
-
-	if !dm.treemapMode {
-		t.Error("expected treemap mode to remain active after global search jump")
-	}
-	if dm.treeMode {
-		t.Error("expected tree mode to remain disabled after global search jump")
-	}
-	if dm.nav.Entry().Path != "root/subdir" {
-		t.Errorf("expected nav to enter subdir, got %q", dm.nav.Entry().Path)
-	}
-
-	_ = dm.View() // trigger treemap block layout
-
-	// After entering subdir, the treemap should show its children (b.py).
-	idx := dm.findTreemapBlockIndex(dm.nav.Entry().GetChild("b.py"))
-	if idx < 0 {
-		t.Fatalf("expected b.py to be visible in treemap blocks after jumping to subdir, got %d blocks", len(dm.treemapBlocks))
-	}
-}
-
-func TestViewModelGlobalSearchJumpsToFile(t *testing.T) {
-	dm := newTestDirModel()
-	dm.Update(ScanFinished{})
-	vm := NewViewModel(dm.nav, dm)
-
-	// Press Ctrl+P to open global search.
-	_, cmd := vm.Update(tea.KeyMsg{Type: tea.KeyCtrlP})
-	if cmd != nil {
-		t.Fatalf("expected Ctrl+P not to return a command, got %v", cmd)
-	}
-	if dm.mode != SEARCH {
-		t.Fatalf("expected SEARCH mode after Ctrl+P, got %v", dm.mode)
-	}
-
-	// Type the query and update the model directly (DirModel handles typing).
-	dm.searchInput.SetValue("b.py")
-	dm.updateSearchQuery()
-
-	// Press Enter to apply the search result.
-	_, cmd = vm.Update(tea.KeyMsg{Type: tea.KeyEnter})
-	if cmd != nil {
-		t.Fatalf("expected Enter not to return a command, got %v", cmd)
-	}
-	if dm.mode != READY {
-		t.Fatalf("expected READY mode after applying search result, got %v", dm.mode)
-	}
-	if dm.nav.Entry().Path != "root" {
-		t.Errorf("expected navigation at root, got %q", dm.nav.Entry().Path)
-	}
-
-	idx := dm.findChildIndex(dm.nav.Entry().GetChild("b.py"))
-	if idx < 0 {
-		t.Fatal("expected b.py to be visible after search jump")
-	}
-	if dm.dirsTable.Cursor() != idx {
-		t.Errorf("expected cursor on b.py (index %d), got %d", idx, dm.dirsTable.Cursor())
-	}
 }
 
 func TestViewModelPreviewQClosesPreviewWithoutQuitting(t *testing.T) {
@@ -521,206 +114,6 @@ func TestViewModelInputQFiltersWithoutQuitting(t *testing.T) {
 	}
 	if got := len(dm.dirsTable.Rows()); got != 0 {
 		t.Fatalf("expected q to be applied to the name filter, got %d rows", got)
-	}
-}
-
-func TestTreemapKeyboardStaysAtTopLevel(t *testing.T) {
-	dm := &DirModel{
-		treemapMode: true,
-		treemapBlocks: []treemapBlock{
-			{entry: &structure.Entry{Path: "a"}, level: 0, topIdx: 0},
-			{entry: &structure.Entry{Path: "a1"}, level: 1, topIdx: 0},
-			{entry: &structure.Entry{Path: "a2"}, level: 1, topIdx: 0},
-			{entry: &structure.Entry{Path: "b"}, level: 0, topIdx: 1},
-			{entry: &structure.Entry{Path: "b1"}, level: 1, topIdx: 1},
-			{entry: &structure.Entry{Path: "c"}, level: 0, topIdx: 2},
-		},
-		treemapSelected: 0,
-	}
-
-	dm.moveTreemapSelection(1)
-	if dm.treemapSelected != 3 {
-		t.Fatalf("expected selection to move to top-level b (index 3), got %d", dm.treemapSelected)
-	}
-
-	dm.moveTreemapSelection(1)
-	if dm.treemapSelected != 5 {
-		t.Fatalf("expected selection to move to top-level c (index 5), got %d", dm.treemapSelected)
-	}
-
-	dm.moveTreemapSelection(-1)
-	if dm.treemapSelected != 3 {
-		t.Fatalf("expected selection to move back to top-level b (index 3), got %d", dm.treemapSelected)
-	}
-
-	// Starting from a nested block, j/k should still move between top-level blocks.
-	dm.treemapSelected = 1 // a1
-	dm.moveTreemapSelection(1)
-	if dm.treemapSelected != 3 {
-		t.Fatalf("expected selection from nested block to jump to top-level b (index 3), got %d", dm.treemapSelected)
-	}
-}
-
-func TestByteIndexesToRuneIndexes(t *testing.T) {
-	// "中" is 3 bytes; "a中b" has byte offsets 0,1,4,5 and rune offsets 0,1,2,3.
-	s := "a中b"
-	byteIdxs := []int{0, 1, 4, 5}
-	want := []int{0, 1, 2, 3}
-
-	got := byteIndexesToRuneIndexes(s, byteIdxs)
-	if len(got) != len(want) {
-		t.Fatalf("expected %v, got %v", want, got)
-	}
-	for i := range want {
-		if got[i] != want[i] {
-			t.Errorf("at index %d: expected %d, got %d", i, want[i], got[i])
-		}
-	}
-}
-
-func TestDirModelGlobalSearchPgDownNoMatches(t *testing.T) {
-	dm := newTestDirModel()
-	dm.Update(ScanFinished{})
-	dm.openGlobalSearch()
-
-	dm.searchInput.SetValue("nonexistent")
-	dm.updateSearchQuery()
-
-	if len(dm.searchMatches) != 0 {
-		t.Fatalf("expected no matches, got %d", len(dm.searchMatches))
-	}
-
-	msg := tea.KeyMsg{Type: tea.KeyPgDown}
-	dm.Update(msg)
-
-	if dm.searchCursor != 0 {
-		t.Errorf("expected cursor to stay at 0 with no matches, got %d", dm.searchCursor)
-	}
-}
-
-func TestDirModelLanguageSelectOverlay(t *testing.T) {
-	dm := newTestDirModel()
-	dm.Update(ScanFinished{})
-
-	if len(dm.languages) != 2 {
-		t.Fatalf("expected 2 languages, got %d", len(dm.languages))
-	}
-	if dm.languages[0] != "Go" || dm.languages[1] != "Python" {
-		t.Errorf("languages = %v, want [Go Python]", dm.languages)
-	}
-
-	dm.Update(tea.KeyMsg{Type: tea.KeyCtrlL})
-	if dm.mode != SELECT_LANG {
-		t.Fatalf("expected SELECT_LANG mode, got %v", dm.mode)
-	}
-	if !dm.selectMode {
-		t.Error("expected selectMode to be true")
-	}
-	if dm.selectIndex != 0 {
-		t.Errorf("expected selectIndex 0, got %d", dm.selectIndex)
-	}
-
-	dm.Update(tea.KeyMsg{Type: tea.KeyDown})
-	if dm.selectIndex != 1 {
-		t.Errorf("after down: selectIndex = %d, want 1", dm.selectIndex)
-	}
-	dm.Update(tea.KeyMsg{Type: tea.KeyUp})
-	if dm.selectIndex != 0 {
-		t.Errorf("after up: selectIndex = %d, want 0", dm.selectIndex)
-	}
-
-	// Boundary at the top.
-	dm.Update(tea.KeyMsg{Type: tea.KeyUp})
-	if dm.selectIndex != 0 {
-		t.Errorf("top boundary: selectIndex = %d, want 0", dm.selectIndex)
-	}
-
-	// Boundary at the bottom.
-	dm.selectIndex = len(dm.languages) - 1
-	dm.Update(tea.KeyMsg{Type: tea.KeyDown})
-	if dm.selectIndex != len(dm.languages)-1 {
-		t.Errorf("bottom boundary: selectIndex = %d, want %d", dm.selectIndex, len(dm.languages)-1)
-	}
-
-	// Toggle selection with space.
-	dm.selectIndex = 0 // Go
-	dm.Update(tea.KeyMsg{Type: tea.KeySpace})
-	if !dm.selectedLangs["Go"] {
-		t.Error("expected Go to be selected")
-	}
-	dm.Update(tea.KeyMsg{Type: tea.KeySpace})
-	if dm.selectedLangs["Go"] {
-		t.Error("expected Go to be deselected")
-	}
-
-	// Enter applies the filter and exits select mode.
-	dm.Update(tea.KeyMsg{Type: tea.KeySpace})
-	dm.Update(tea.KeyMsg{Type: tea.KeyEnter})
-	if dm.mode != READY {
-		t.Errorf("expected READY after enter, got %v", dm.mode)
-	}
-	if !dm.selectedLangs["Go"] {
-		t.Error("expected Go to remain selected")
-	}
-	if len(dm.dirsTable.Rows()) != 2 {
-		t.Errorf("expected 2 Go rows, got %d", len(dm.dirsTable.Rows()))
-	}
-
-	// Escape closes language select without applying the latest selection.
-	dm.Update(tea.KeyMsg{Type: tea.KeyCtrlL})
-	dm.selectIndex = 1 // Python
-	dm.Update(tea.KeyMsg{Type: tea.KeySpace})
-	dm.Update(tea.KeyMsg{Type: tea.KeyEsc})
-	if dm.mode != READY {
-		t.Errorf("expected READY after escape, got %v", dm.mode)
-	}
-	if dm.selectMode {
-		t.Error("expected selectMode to be false")
-	}
-	// The previously applied Go filter is still in effect.
-	if len(dm.dirsTable.Rows()) != 2 {
-		t.Errorf("expected 2 rows after escape, got %d", len(dm.dirsTable.Rows()))
-	}
-}
-
-func TestDirModelLanguageSelectQClosesOverlay(t *testing.T) {
-	dm := newTestDirModel()
-	dm.Update(ScanFinished{})
-
-	dm.Update(tea.KeyMsg{Type: tea.KeyCtrlL})
-	if dm.mode != SELECT_LANG {
-		t.Fatalf("expected SELECT_LANG mode, got %v", dm.mode)
-	}
-
-	dm.Update(tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'q'}})
-	if dm.mode != READY {
-		t.Errorf("expected q to close language select and return to READY, got %v", dm.mode)
-	}
-	if dm.selectMode {
-		t.Error("expected selectMode to be false after q")
-	}
-}
-
-func TestDirModelLanguageSelectEscRevertsChanges(t *testing.T) {
-	dm := newTestDirModel()
-	dm.Update(ScanFinished{})
-
-	dm.Update(tea.KeyMsg{Type: tea.KeyCtrlL})
-	dm.Update(tea.KeyMsg{Type: tea.KeySpace}) // select Go
-	if !dm.selectedLangs["Go"] {
-		t.Fatal("expected Go to be selected before escape")
-	}
-
-	// Esc should revert to the empty selection state from before the overlay opened.
-	dm.Update(tea.KeyMsg{Type: tea.KeyEsc})
-	if dm.mode != READY {
-		t.Errorf("expected READY after escape, got %v", dm.mode)
-	}
-	if dm.selectedLangs["Go"] {
-		t.Error("expected Go selection to be reverted after escape")
-	}
-	if len(dm.dirsTable.Rows()) != 3 {
-		t.Errorf("expected all 3 rows after reverting selection, got %d", len(dm.dirsTable.Rows()))
 	}
 }
 
@@ -915,189 +308,6 @@ func TestDirModelModeToggles(t *testing.T) {
 	}
 }
 
-func TestDirModelMouseHelpers(t *testing.T) {
-	dm := newTestDirModel()
-
-	dm.overlayBounds = overlayBounds{kind: "preview", x: 10, y: 5, w: 40, h: 20}
-	cases := []struct {
-		x, y int
-		want bool
-	}{
-		{10, 5, true},
-		{49, 24, true},
-		{9, 5, false},
-		{50, 5, false},
-		{10, 4, false},
-		{10, 25, false},
-	}
-	for _, tc := range cases {
-		got := dm.isInsideOverlay(tc.x, tc.y)
-		if got != tc.want {
-			t.Errorf("isInsideOverlay(%d,%d) = %v, want %v", tc.x, tc.y, got, tc.want)
-		}
-		if !dm.isInsidePreviewBox(tc.x, tc.y) && tc.want {
-			t.Errorf("isInsidePreviewBox(%d,%d) should be true", tc.x, tc.y)
-		}
-	}
-
-	dm.overlayBounds.kind = "chart"
-	if !dm.isInsideChartBox(15, 10) {
-		t.Error("expected inside chart box")
-	}
-	if dm.isInsidePreviewBox(15, 10) {
-		t.Error("expected not inside preview box")
-	}
-
-	dm.overlayBounds.kind = "langselect"
-	if !dm.isInsideLangSelectBox(15, 10) {
-		t.Error("expected inside language select box")
-	}
-}
-
-func TestDirModelLangSelectIndexAtY(t *testing.T) {
-	dm := newTestDirModel()
-	dm.languages = []string{"A", "B", "C", "D", "E"}
-
-	t.Run("no scroll offset", func(t *testing.T) {
-		dm.overlayBounds = overlayBounds{kind: "langselect", x: 0, y: 5, w: 20, h: 10, langStart: 0, langEnd: 3}
-		if got := dm.langSelectIndexAtY(5 + 1); got != -1 { // title line
-			t.Errorf("title line = %d, want -1", got)
-		}
-		if got := dm.langSelectIndexAtY(5 + 3); got != 0 { // first list item
-			t.Errorf("first item = %d, want 0", got)
-		}
-		if got := dm.langSelectIndexAtY(5 + 4); got != 1 {
-			t.Errorf("second item = %d, want 1", got)
-		}
-		if got := dm.langSelectIndexAtY(5 + 6); got != -1 { // past last visible item
-			t.Errorf("past end = %d, want -1", got)
-		}
-	})
-
-	t.Run("with scroll offset", func(t *testing.T) {
-		dm.overlayBounds = overlayBounds{kind: "langselect", x: 0, y: 5, w: 20, h: 10, langStart: 2, langEnd: 5}
-		if got := dm.langSelectIndexAtY(5 + 2); got != -1 { // "..." line
-			t.Errorf("ellipsis line = %d, want -1", got)
-		}
-		if got := dm.langSelectIndexAtY(5 + 4); got != 2 { // first visible item
-			t.Errorf("first visible item = %d, want 2", got)
-		}
-		if got := dm.langSelectIndexAtY(5 + 6); got != 4 { // last visible item
-			t.Errorf("last visible item = %d, want 4", got)
-		}
-	})
-}
-
-func TestDirModelTableRowAtY(t *testing.T) {
-	dm := newTestDirModel()
-	dm.Update(ScanFinished{})
-	dm.updateSize(120, 24)
-	dm.lastTableView = dm.dirsTable.View()
-
-	cursorLine := dm.findCursorLineInView(dm.lastTableView)
-	if cursorLine < 0 {
-		t.Fatalf("cursor line not found in view:\n%s", dm.lastTableView)
-	}
-
-	y := tableHeaderHeight + cursorLine
-	got := dm.tableRowAtY(y)
-	want := dm.dirsTable.Cursor()
-	if got != want {
-		t.Errorf("tableRowAtY(%d) = %d, want %d", y, got, want)
-	}
-
-	if dm.tableRowAtY(0) != -1 {
-		t.Error("expected header Y to return -1")
-	}
-	if dm.tableRowAtY(1000) != -1 {
-		t.Error("expected out-of-range Y to return -1")
-	}
-}
-
-func TestDirModelFindCursorLineInView(t *testing.T) {
-	dm := newTestDirModel()
-	dm.Update(ScanFinished{})
-	dm.updateSize(120, 24)
-
-	view := dm.dirsTable.View()
-	line := dm.findCursorLineInView(view)
-	if line < 0 {
-		t.Fatalf("findCursorLineInView returned %d", line)
-	}
-
-	lines := strings.Split(view, "\n")
-	if line >= len(lines)-tableHeaderHeight {
-		t.Errorf("cursor line %d out of range for view with %d data lines", line, len(lines)-tableHeaderHeight)
-	}
-}
-
-func TestTreemapColorModeToggle(t *testing.T) {
-	dm := newTestDirModel()
-	dm.Update(ScanFinished{})
-	dm.treemapMode = true
-	dm.width = 100
-	dm.height = 30
-	dm.updateTableData()
-
-	if dm.treemapColorByLang {
-		t.Fatal("expected default directory color mode")
-	}
-
-	dm.Update(tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'c'}})
-	if !dm.treemapColorByLang {
-		t.Fatal("expected language color mode after pressing c")
-	}
-	if !strings.Contains(dm.View(), "Languages") {
-		t.Fatal("expected legend to auto-show in language color mode")
-	}
-
-	dm.Update(tea.KeyMsg{Type: tea.KeyRunes, Runes: []rune{'c'}})
-	if dm.treemapColorByLang {
-		t.Fatal("expected directory color mode after second c")
-	}
-	if strings.Contains(dm.View(), "Languages") {
-		t.Fatal("expected legend to hide in directory color mode")
-	}
-}
-
-func TestTreemapLegendAutoShow(t *testing.T) {
-	dm := newTestDirModel()
-	dm.Update(ScanFinished{})
-	dm.treemapMode = true
-	dm.treemapColorByLang = true
-	dm.width = 100
-	dm.height = 30
-	dm.updateTableData()
-
-	view := dm.View()
-	if !strings.Contains(view, "Languages") {
-		t.Fatalf("expected legend to auto-show in language color mode, got:\n%s", view)
-	}
-}
-
-func TestTreemapViewHeight(t *testing.T) {
-	dm := newTestDirModel()
-	dm.Update(ScanFinished{})
-	dm.treemapMode = true
-	dm.treemapColorByLang = true
-	dm.width = 100
-	dm.height = 30
-	dm.updateTableData()
-
-	view := dm.View()
-	got := lipgloss.Height(view)
-	if got != dm.height {
-		t.Fatalf("expected view height %d, got %d", dm.height, got)
-	}
-}
-
-func sccProviderInfo() provider.Info {
-	return provider.Info{
-		Name:         "scc",
-		Capabilities: provider.CapLines | provider.CapComplexity,
-	}
-}
-
 func TestNewDirModel_DynamicColumns(t *testing.T) {
 	root := structure.NewDirEntry("root")
 	root.AggregateStats()
@@ -1120,297 +330,5 @@ func TestNewDirModel_DynamicColumns(t *testing.T) {
 		if got[i] != want[i] {
 			t.Errorf("column %d: expected %q, got %q", i, want[i], got[i])
 		}
-	}
-}
-
-func TestVisibleColumns_Wide(t *testing.T) {
-	dm := NewDirModel(NewCodeNavigation(structure.NewTree(structure.NewDirEntry("root"))), sccProviderInfo(), false, false)
-	dm.width = 120
-	dm.sortState = SortState{Key: SortByTotal, Desc: true}
-
-	cols := dm.visibleColumns()
-	if len(cols) != len(dm.columns) {
-		t.Errorf("expected all %d columns at width 120, got %d", len(dm.columns), len(cols))
-	}
-}
-
-func TestVisibleColumns_NarrowHidesOptional(t *testing.T) {
-	dm := NewDirModel(NewCodeNavigation(structure.NewTree(structure.NewDirEntry("root"))), sccProviderInfo(), false, false)
-	dm.width = 50
-	dm.sortState = SortState{Key: SortByTotal, Desc: true}
-
-	cols := dm.visibleColumns()
-	for _, c := range cols {
-		switch c.SortKey {
-		case SortByComplexity, SortByLanguages, SortByComments, SortByBlanks:
-			t.Errorf("expected %q to be hidden at width 50, but it was visible", c.SortKey)
-		}
-	}
-}
-
-func TestVisibleColumns_ActiveSortShownDespiteNarrow(t *testing.T) {
-	dm := NewDirModel(NewCodeNavigation(structure.NewTree(structure.NewDirEntry("root"))), sccProviderInfo(), false, false)
-	dm.width = 50
-	dm.sortState = SortState{Key: SortByComplexity, Desc: true}
-
-	cols := dm.visibleColumns()
-	found := false
-	for _, c := range cols {
-		if c.SortKey == SortByComplexity {
-			found = true
-		}
-	}
-	if !found {
-		t.Errorf("expected active sort column %q to remain visible at width 50", SortByComplexity)
-	}
-}
-
-func TestBuildRow_WithComplexity(t *testing.T) {
-	root := structure.NewDirEntry("root")
-	root.AddChild(structure.NewFileEntry("root/a.go", map[string]structure.CodeStats{
-		"Go": {Code: 20, Comments: 5, Blanks: 5, Complexity: 7},
-	}))
-	root.AggregateStats()
-
-	dm := NewDirModel(NewCodeNavigation(structure.NewTree(root)), sccProviderInfo(), false, false)
-	entry := root.Child[0]
-	stats := entry.TotalStats
-
-	row := dm.buildRow(dm.columns, entry, "a.go", "Go", stats, 100.0)
-	if len(row) != len(dm.columns) {
-		t.Fatalf("expected row length %d, got %d", len(dm.columns), len(row))
-	}
-
-	want := map[SortKey]string{
-		SortByName:       "a.go",
-		SortByLanguages:  "Go",
-		SortByCode:       "20",
-		SortByComments:   "5",
-		SortByBlanks:     "5",
-		SortByTotal:      "30",
-		SortByPercent:    "100.00 %",
-		SortByComplexity: "7",
-	}
-	for i, c := range dm.columns {
-		if wantVal, ok := want[c.SortKey]; ok && row[i] != wantVal {
-			t.Errorf("column %q: expected %q, got %q", c.SortKey, wantVal, row[i])
-		}
-	}
-}
-
-func TestBuildChildComparator_Complexity(t *testing.T) {
-	root := structure.NewDirEntry("root")
-	root.AddChild(structure.NewFileEntry("root/a.go", map[string]structure.CodeStats{
-		"Go": {Code: 10, Complexity: 5},
-	}))
-	root.AddChild(structure.NewFileEntry("root/b.go", map[string]structure.CodeStats{
-		"Go": {Code: 10, Complexity: 3},
-	}))
-	root.AggregateStats()
-
-	dm := NewDirModel(NewCodeNavigation(structure.NewTree(root)), sccProviderInfo(), false, false)
-
-	dm.sortState = SortState{Key: SortByComplexity, Desc: true}
-	cmp := dm.buildChildComparator()
-	if cmp(root.Child[0], root.Child[1]) >= 0 {
-		t.Error("expected a.go (complexity 5) to come before b.go (complexity 3)")
-	}
-}
-
-func TestCycleTreemapSize_WithSCC(t *testing.T) {
-	dm := NewDirModel(
-		NewCodeNavigation(structure.NewTree(structure.NewDirEntry("root"))),
-		sccProviderInfo(),
-		false,
-		true,
-	)
-
-	order := []SortKey{SortByTotal, SortByComplexity}
-	for i := 0; i < len(order)*2; i++ {
-		expected := order[i%len(order)]
-		if dm.treemapSizeKey != expected {
-			t.Errorf("cycle %d: expected size key %q, got %q", i, expected, dm.treemapSizeKey)
-		}
-		dm.cycleTreemapSize()
-	}
-}
-
-func TestCycleTreemapSize_WithTokei(t *testing.T) {
-	dm := NewDirModel(
-		NewCodeNavigation(structure.NewTree(structure.NewDirEntry("root"))),
-		provider.Info{Name: "tokei", Capabilities: provider.CapLines},
-		false,
-		true,
-	)
-
-	for i := 0; i < 3; i++ {
-		dm.cycleTreemapSize()
-		if dm.treemapSizeKey != SortByTotal {
-			t.Errorf("tokei provider should only support Total size, got %q", dm.treemapSizeKey)
-		}
-	}
-}
-
-func TestTreemapSizeFunc(t *testing.T) {
-	root := structure.NewDirEntry("root")
-	root.AddChild(structure.NewFileEntry("root/a.go", map[string]structure.CodeStats{
-		"Go": {Code: 10, Comments: 2, Blanks: 3, Complexity: 5},
-	}))
-	root.AggregateStats()
-
-	dm := NewDirModel(NewCodeNavigation(structure.NewTree(root)), sccProviderInfo(), false, true)
-
-	dm.treemapSizeKey = SortByTotal
-	if got := dm.treemapSizeFunc()(root.Child[0]); got != 15 {
-		t.Errorf("Total size: expected 15, got %d", got)
-	}
-
-	dm.treemapSizeKey = SortByComplexity
-	if got := dm.treemapSizeFunc()(root.Child[0]); got != 5 {
-		t.Errorf("Complexity size: expected 5, got %d", got)
-	}
-}
-
-func TestPercentColumn_ContextAware(t *testing.T) {
-	root := structure.NewDirEntry("root")
-	root.AddChild(structure.NewFileEntry("root/a.go", map[string]structure.CodeStats{
-		"Go": {Code: 10, Comments: 5, Blanks: 5, Complexity: 5},
-	}))
-	root.AddChild(structure.NewFileEntry("root/b.go", map[string]structure.CodeStats{
-		"Go": {Code: 30, Comments: 10, Blanks: 10, Complexity: 15},
-	}))
-	root.AggregateStats()
-
-	dm := NewDirModel(NewCodeNavigation(structure.NewTree(root)), sccProviderInfo(), false, false)
-	dm.width = 120
-	dm.height = 30
-
-	tests := []struct {
-		key      SortKey
-		row0     string
-		row1     string
-	}{
-		{SortByTotal, "71.43 %", "28.57 %"},
-		{SortByComplexity, "75.00 %", "25.00 %"},
-	}
-
-	for _, tt := range tests {
-		t.Run(string(tt.key), func(t *testing.T) {
-			dm.sortState = SortState{Key: tt.key, Desc: true}
-			dm.updateTableData()
-			rows := dm.dirsTable.Rows()
-			if len(rows) < 2 {
-				t.Fatalf("expected 2 rows, got %d", len(rows))
-			}
-			// Find the percent column index.
-			percentIdx := -1
-			for i, c := range dm.columns {
-				if c.SortKey == SortByPercent {
-					percentIdx = i
-					break
-				}
-			}
-			if percentIdx < 0 {
-				t.Fatal("percent column not found")
-			}
-			if got := rows[0][percentIdx]; got != tt.row0 {
-				t.Errorf("row0 percent: expected %s, got %s", tt.row0, got)
-			}
-			if got := rows[1][percentIdx]; got != tt.row1 {
-				t.Errorf("row1 percent: expected %s, got %s", tt.row1, got)
-			}
-		})
-	}
-}
-
-func TestDirModelHandleHeaderClick(t *testing.T) {
-	dm := newTestDirModel()
-	dm.Update(ScanFinished{})
-	dm.updateSize(120, 24)
-
-	cols := dm.dirsTable.Columns()
-	if len(cols) < 8 {
-		t.Fatalf("expected at least 8 columns, got %d", len(cols))
-	}
-
-	midX := func(idx int) int {
-		startX := 0
-		for i := 0; i < idx; i++ {
-			startX += cols[i].Width
-		}
-		return startX + cols[idx].Width/2
-	}
-
-	t.Run("click sortable column changes key", func(t *testing.T) {
-		dm.sortState = SortState{Key: SortByTotal, Desc: true}
-		if !dm.handleHeaderClick(midX(2)) {
-			t.Fatal("expected Name header click to be handled")
-		}
-		if dm.sortState.Key != SortByName {
-			t.Errorf("expected sort key %q, got %q", SortByName, dm.sortState.Key)
-		}
-		if dm.sortState.Desc != false {
-			t.Errorf("expected Name to default ascending, got Desc=%v", dm.sortState.Desc)
-		}
-	})
-
-	t.Run("click current sort column toggles order", func(t *testing.T) {
-		dm.sortState = SortState{Key: SortByName, Desc: false}
-		if !dm.handleHeaderClick(midX(2)) {
-			t.Fatal("expected Name header click to be handled")
-		}
-		if dm.sortState.Key != SortByName {
-			t.Errorf("expected sort key to stay %q, got %q", SortByName, dm.sortState.Key)
-		}
-		if dm.sortState.Desc != true {
-			t.Errorf("expected Desc to toggle to true, got %v", dm.sortState.Desc)
-		}
-	})
-
-	t.Run("click non-sortable column does nothing", func(t *testing.T) {
-		dm.sortState = SortState{Key: SortByTotal, Desc: true}
-		if dm.handleHeaderClick(midX(0)) {
-			t.Error("expected icon header click to be ignored")
-		}
-		if dm.sortState.Key != SortByTotal {
-			t.Errorf("expected sort key to remain %q, got %q", SortByTotal, dm.sortState.Key)
-		}
-	})
-
-	t.Run("click outside columns does nothing", func(t *testing.T) {
-		dm.sortState = SortState{Key: SortByTotal, Desc: true}
-		totalWidth := 0
-		for _, c := range cols {
-			totalWidth += c.Width
-		}
-		if dm.handleHeaderClick(totalWidth + 10) {
-			t.Error("expected click outside columns to be ignored")
-		}
-	})
-}
-
-func TestDirModelHandleTableMouseRoutesHeaderClicks(t *testing.T) {
-	dm := newTestDirModel()
-	dm.Update(ScanFinished{})
-	dm.updateSize(120, 24)
-
-	cols := dm.dirsTable.Columns()
-	nameX := cols[0].Width + cols[1].Width + cols[2].Width/2
-
-	dm.sortState = SortState{Key: SortByTotal, Desc: true}
-	row, clicks, handled := dm.handleTableMouse(tea.MouseMsg{
-		Button: tea.MouseButtonLeft,
-		Action: tea.MouseActionPress,
-		X:      nameX,
-		Y:      0,
-	})
-	if !handled {
-		t.Fatal("expected header click to be handled")
-	}
-	if row != -1 || clicks != 0 {
-		t.Errorf("expected row=-1, clicks=0, got row=%d, clicks=%d", row, clicks)
-	}
-	if dm.sortState.Key != SortByName {
-		t.Errorf("expected sort key to change to %q, got %q", SortByName, dm.sortState.Key)
 	}
 }
